+++ example/lib/data/source/fitter_sample_source.dart
@@ -18,4 +18,3 @@ class SampleFittorSource extends FitUrls implements FitterSampleRepo {
     throw UnimplementedError();
   }
 }
-